// integrations/gohighlevel/src/misc/custom-schemas.ts
export const makeApiCallInputSchema = z.object({
  endpoint: z.string().describe('The API endpoint to call'),
  method: z.string().describe("The HTTP method to use ['GET', 'POST', 'PUT', 'DELETE']"),
  data: z.string().optional().describe('The data to send with the request as a string JSON object.'),
  params: z.string().optional().describe('The params to send with the request as a string JSON object if required.')
})